[#2] Refactor: 예약 버튼 모달 내부로 수정
기존 리스트 아이템에 적용했던 예약 버튼 기능을 모달 내부로 옮기도록 하였습니다.
