+++ src/components/ProductCard.tsx
@@ -1,17 +1,5 @@
-import {
-  Card,
-  CardBody,
-  Image,
-  Stack,
-  Text,
-  Heading,
-  Flex,
-  Button,
-  useToast,
-} from '@chakra-ui/react';
+import { Card, CardBody, Image, Stack, Text, Heading, Flex } from '@chakra-ui/react';
 import { useState } from 'react';
-
-import { RESERVE_LIST_KEY } from '@/constants';
 
 import { ProductModal } from './ProductModal';
 
@@ -19,29 +7,9 @@ export const ProductCard = (props: Product) => {
   const { idx, name, mainImage, description, price, spaceCategory } = props;
 
   const [isOpen, setIsOpen] = useState<boolean>(false);
-  const toast = useToast();
 
   const handleModal = () => {
     setIsOpen((prevIsOpen) => !prevIsOpen);
-  };
-
-  const handleReserve = (e: React.MouseEvent<HTMLButtonElement>) => {
-    e.stopPropagation();
-    const reserveList = localStorage.getItem(RESERVE_LIST_KEY)?.split(' ') || [];
-    if (reserveList.includes(idx + '')) {
-      toast({
-        title: '이미 예약하신 상품입니다.',
-        status: 'error',
-        duration: 2000,
-      });
-      return;
-    }
-    toast({
-      title: '예약이 완료되었습니다.',
-      status: 'info',
-      duration: 2000,
-    });
-    localStorage.setItem(RESERVE_LIST_KEY, `${reserveList.join(' ')} ${idx}`);
   };
 
   return (
@@ -65,9 +33,6 @@ export const ProductCard = (props: Product) => {
                 {spaceCategory}
               </Text>
             </Flex>
-            <Button colorScheme='blue' onClick={handleReserve}>
-              예약
-            </Button>
           </Stack>
         </CardBody>
       </Card>
